Fixed Calculation of Closest Point
currentlyCloser is now a CGFloat instead of an NSInteger. See issue #58.

+++ Classes/BEMSimpleLineGraphView.m
@@ -23,7 +23,7 @@
     
     /// The closest point to the touch point
     BEMCircle *closestDot;
-    NSInteger currentlyCloser;
+    CGFloat currentlyCloser;
     
     /// All of the X-Axis Values
     NSMutableArray *xAxisValues;
